[Update] 차량 추가관련 코드 수정하였습니다.

+++ src/main/java/xyz/wavey/vehicleservice/vehicle/model/Vehicle.java
@@ -61,7 +61,7 @@ public class Vehicle extends BaseTimeEntity {
     private String smartKey;
 
     @Column(name = "last_zone", nullable = false)
-    private Integer lastZone;
+    private Long lastZone;
 
     @Column(name = "wash_time", nullable = false)
     private Date washTime;
