+++ src/main/java/xyz/wavey/vehicleservice/vehicle/vo/RequestVehicle.java
@@ -5,7 +5,6 @@ import java.util.Date;
 import java.util.Map;
 import lombok.Getter;
 import lombok.Setter;
-import xyz.wavey.vehicleservice.frame.model.Frame;
 
 @Getter
 @Setter
@@ -18,7 +17,7 @@ public class RequestVehicle {
   private Boolean available;
   private Integer charge;
   private String image;
-  private Integer lastZone;
+  private Long lastZone;
   private String smartKey;
   private Long frameId;
   private Date washTime;
